further cleaned the programmatically cleaned catalog
We're at the point where I can start trying to parse it, but further
cleaning may be required (as I learn more about the necessary
constraints for consistent parsing).

While manually cleaning it (because I thought it wasn't doable
programatically), I realized that it very well *may* be possible to
automate. So if more of the same type of manual parsing is required
in the future, we may be able to automate it (at least for the COMP
subject).